refactor: 💡 Convert SQFormMaskedTextField to TS
✅ Closes: #379
