+++ package.json
@@ -124,6 +124,7 @@
     "@types/jest": "^26.0.23",
     "@types/react": "^17.0.5",
     "@types/react-dom": "^17.0.4",
+    "@types/react-text-mask": "^5.4.8",
     "@types/testing-library__jest-dom": "^5.9.5",
     "@types/yup": "^0.29.11",
     "@typescript-eslint/eslint-plugin": "^4.23.0",
