+++ src/types/MaskTypes.ts
@@ -1,0 +1,3 @@
+import {maskArray} from 'react-text-mask';
+
+export type maskProp = maskArray | ((value: string) => maskArray);
